Roadmap revision: no jsx_proplist format
Upon some reflection (and discussion with @mrallen1), I’ve decided not to deal with the so-called `jsx_proplist` format I had suggested as an avenue for jsn to support jsx/jsone-decoded `proplist` objects.

This is little more than confusion. Clients should simply avoid using proplists if they are dealing with empty objects. The `map` format is a strongly encouraged alternative that will become the default format in v2.0.0.