get_nth object array fixes
I noticed in looking at #24 that there was an issue with `get_nth` where it would improperly handle arrays of objects. I showed 3 examples in a comment on #25:

```erlang
1> jsn:get({<<"foo">>, 1}, jsn:new([{<<"foo.bar">>, 1}], [{format, struct}])).
{<<"bar">>,1} % should be undefined
2> Object3Struct = {struct, [{<<"foo">>, [{struct, [{<<"bar">>, <<"baz1">>}]},
2>                                        {struct, [{<<"bar">>, <<"baz2">>}]}]}]}.
3> jsn:get({<<"foo">>, first, <<"bar">>}, Object3Struct).
undefined % should be <<"baz1">>
3> jsn:get({<<"foo">>, last, <<"bar">>}, Object3Struct).
undefined % should be <<"baz2">>
```

This change corrects the cases above, and generally improves and unifies the logic of `get_nth`. I have also added tests to the get_nth suite for arrays with JSON object and sub-array elements.