Patch jsn:delete to delete objects at end of path
Resolves #42.

**Problem**

when calling `jsn:delete/2` on a path to an object where the value is itself an object, `jsn` doesn't delete the object due to incorrect guards in the handling of deletion for nested paths: the guard does not work when there *isn't* a nested path element to walk into, leaving the leaf object.

**Solution**

Fix the guard and update the tests.